Update checking of the kickstart files MD5 checksusm.
This updates the CI script used to verify the MD5 checksums of the
kickstart files as being correct in the documentation to use the new
locations for the respective checksums.

This is more involved than a simple path update because the existing
script assumes that both checksums are listed in the same file, which is
no longer the case.

Any future updates that move the location of the checksums just need to
modify the lines in tests/installer/checksums.sh that start with
`check_checksum` to point to the correct files.