Add rudimentary disk size detection to system-info.sh.
This adds really basic detection of total disk size to the
system-info.sh script. This detection is _wildly_ inaccurate in many
cases, but it's not realistically possible to make it much better than
what's added here without making the script exponentially more
complicated (there are just way too many edge cases and special
conditions to worry about).

This adds the following keys to the output of the script:

* NETDATA_TOTAL_DISK_SIZE: This indicates the total detected disk size
  in bytes.
* NETDATA_DISK_DETECTION: This indicates what detection method was used
  for determining the total disk size. It may return either `df` (which
  works almost everywhere but is wildly inaccurate) or `sysfs` (which is
  linux specific, but is 100% accurate for a majority of cases).

On most platforms, this parses the output of the `df` command, limiting
only to filesystems that are used on fixed disks, which provides
reasonably accurate results in most trvial cases, but kind of falls
apart the moment people are doing anything remotely complicated in terms
of storage (like using a logical volume manager or under-comitting their
disk space).

On Linux, it will preferentially try to parse the info out of
`/sys/block`, filtering on device major numbers that are actually used
for fixed disks and excluding devices that are indicated to be
removable. This provides a very accurate result if the system does not
use removable media as primary storage, but requires that the user who
runs the script can read the contents of `/sys/block`.