Change build dependency for RPMs fron `git` to `git-core`.
We don't actually need everything the regular `git` package pulls in,
just the `git` command and a handful of built-in sub-commands, so we can
just depend on `git-core` during the build instead of the full package.

This causes no issues for existing build workflows, as `git` pulls in
`git-core`. However, it will allow us to make the new Docker-based
package building workflow a bit more efficient as we can decrease the
size of the package builder images.