Don't push a commit for i386 RPM builds.
We're not actually building any i386 RPM packages, and it maes
essentially no sense to do so as the various major RPM-based distros
don't even really support 32-bit x86 installs anymore.

This will actually speed up the builds for the other package types
because they won't be competing for CI resources. It will also make the
commit history tidier in the repo.