Add Clear Linux support to install-required-packages.sh
This adds preliminary support for Clear Linux to the
install-required-packages script, which allows users to use the regular
kickstart install script there.

As of right now, this has a significant limitation in that Clear Linux
does not provide a package for libJudy, meaning that users will not be
able to use the dbengine.