Fixed reconnect issues on the ACLK. (#8163)
Not all of the callbacks generated by LWS have their return codes checked, so there are callback
kinds that cannot ask for the WSI to be destroyed. Connection failures at different phases in the
establishment phase trigger different callback from different points. If the connection fails early
enough then no LWS_CALLBACK_WSI_DESTROY is generated and so we did not trigger our own callback to
inform the upper layer.

When the connection is torn down via callback that does not lead to the WSI_DESTROY the WSI is
internally deallocated after the callback - i.e. the pointer stops being valid at the end of the
callback. We handle this by clearing the pointer and crossing our fingers that there is no memory
leak as LWS gives us no visibility into this and if we call lws_close_free_wsi() twice on the same
pointer it can cause a segmentation fault.

Ultimately this fix is problematic and we may have to fork libwebsockets to fix this poor aspect of
their API design.

This was tested in docker-compose by taking he broker/agent machines up/down in each combination.
Merging with a single review as I need to rebase the feature branch to get valgrind into the
testing environment.