Adding valgrind to container, fixing build error.
The build error is weird - the last PR built ok which suggests that either our Makefile doesn't
capture dependencies properly (so that removing the field from the header file did not trigger
a remake of the .c file), or something is wrong in the incremental build inside the container
(i.e. the .Po files are out of date).

There are some broken arch mirrors right now - worked around in the clean install Dockerfile.

Using `build_external/bin/clean-install arch extras` will put valgrind into the agent container
and use it. This is now ready for testing inside the compose.