Valgrind testing of agent during ACLK reconnection.
It was not so clean to get valgrind into the compose as the entrypoint is overwritten with an
explicit script. This allows valgrind to be used by specifying the alternative compose.

Found memory leaks: I will create a [WIP] PR for this branch so that I can post the leak
reports on it.