Consolidate and clean up artifact validation.
This consolidates the two separate 'Artifacts Validation' stages we have
in Travis into a single stage,a s well as fixing an error in the
handling of the lifecycle checks.

Overall, this should improve build times for Travis CI checks due to a
deduplication of certain lifecycle checks and an overall increase in the
parallelizability of the jobs within the stage.