Switch to triggering on tags instead of commits.
This switches the package build process to use tags as a trigger instead
of requiring specially formatted commit messages. This makes it more
robust and significantly reduces the total amount of code in the PR.

This includes a change to the Travis CI configuration to push the
appropriate tag on nightly builds.