Add option to installer to require cloud to build.
This adds an option to the installer that causes the install to
explicitly fail if we cna't build the agent with netdata Cloud
functionality.

This is primarily intended for testing the Netdata Cloud parts of the
installer in CI, but is also potentially of interest to users who want
to make sure they have this functionality.