Stable stress-testing
The main thread dumps a 2MB payload for writing to the broker every time the write queue is empty.
Currently we are thrashing the LWS event loop to see what it does, and interleaving with the
mosquitto event loop. It take between 40-60s for the payload to arrive at the inspection point.
The write-queue in the agent stays full for this period.