Some proper performance fixes.
Message disassembly before transmission. It looks as if the MQTT broker reassembles the messages from the
stream of independent websockets messages. This would make sense.

There is a massive delay in the lws library between asking for the writable callback and actually receiving
it. This is to do with the size of the SND_BUFFER option on the socket. This fix is a bit CPU heavy as it
cancels the poll() to force a reattempt.

This needs to be properly integrated into our poll-loop.