Added Docker build arg to pass extra options to installer.
This adds a build argument to the main Dockerfile to allow passing extra
options to the netdata-installer script during the build process.

The value specified in the `EXTRA_INSTALL_OPTS` argument wil be added
verbatm to the netdata-installer.sh script. The arguments
`--dont-start-it` and `--dont-wait` are still implicitly passed to the
script as those are required to get a correctly functioning Docker
image.