Explicitly include libffi on ArchLinux CI tests. (#8699)
The `make` command depends on it but does not pull it in as a
dependency. This isn't likely to affect user systems in most cases as
they will almost certainly have Python installed, which also depends on
libffi and pulls it in correctly.

This is a workaround for an upstream bug in Arch.