Added missing path
In the command line given to edit the file `charts.d.conf`, the path to the script `edit-config` is not given so if you type it or paste it that way it won't work. So I think the easiest way to fix it would be to add the path in the command line given in the documentation.