Don't overwrite netdata.conf on update on static installs.
The current code overwrites netdata.conf when updating a static install
because we include a copy of the default empty netdata.conf in the
package itself.

This removes that file from the generated package, but does not modify
the install-time logic which attempts to create the file if it does not
already exist (which appears to be working correctly.