Combined sender buffer state into a single location.
Moved the sender_state into the host, started moving all related entries from the host inside it.
Wrapped up access to the build (BUFFER) and buffer (cbuffer) and fixed the locking. Cleaned up
the old use of locking the buffer to separate out the points in the code that actually touch the
buffer being built. Fixed assumptions about relationship between the thread and the buffer.