Remove knatsakis and ncmans from CODEOWNERS for the agent.
Based on discussion within the SRE team. They are not directly involved
anymore with the agent components maintained by the SRE team, so there's
no point in pestering them all the time with review requests when the
rest of the SRE team is suficient for a complete review.

In the event that a change warrants their attention, they'll be
explicitly added as reviewers by hand.