Liveness on receiver sockets
Dropped the parser_next code for a local socket reader inside the streaming receiver. Allows update of timestamps on
message reception. When we hit multiple connections on the same host we can see if the old connection is still alive.

Currently possible to crash as the locking around the receivers is not implemented yet.