Fill in gap data that has been replicated.
There is a weird bug to track down tomorrow. When the host is scanned to check the gap length the code looks at
rrdset_last_entry_t for each chart in the host. After the first chart is checked the time stamps on all the following
charts jump forward by 26s. This causes the gap length to be calculated incorrectly and so the replication request
doesn't retrieve enough data to fill the gap.

2020-06-04 21:22:11: netdata INFO  : STREAM_RECEIVER[7d1e3909e8eb,[192.168.16.2]:55540] : (1507@database/r:rrdhost_last_en): Chart system.load last 1591298500 global 0 gap=31
2020-06-04 21:22:11: netdata INFO  : STREAM_RECEIVER[7d1e3909e8eb,[192.168.16.2]:55540] : (1507@database/r:rrdhost_last_en): Chart system.idlejitter last 1591298526 global 1591298500 gap=5