Sender-side replication flags (thread-safe)
The receiver-side does not care about thread safety for the replication flag (as only the receiver thread will access
it). On the sender-side the flag is accessed by both the collector thread and the sender thread so it needs to be
thread-safe. The existing chart flags cannot be used as when autotools did not detect the atomic operations they
fail to be thread-safe.

Thread-safety requires a new mutex in the RRDSET structure. This cannot be placed in the unused storage as the size
ratio between size_t and the mutex structure might not be platform independent... After checking the various memory
modes and the use of mmap on the RRDSET structure I believe it is safe to add volatile data to the end of the
structure as long as it is initialized in rrdset_create_custom. This does not change the magic signature in the
structure *or the placement of any existing field*.