Added better checks for existing installs to the kickstart scripts.
This adds more robust checks to our Kickstart install scripts to prevent
users from accidentally installing Netdata over top of an existing
install (which can lead to all kinds of problematic situations).

The logic for these checks works as follows:

* See if `netdata` is located somewhere in `$PATH` (using `command -v`),
  and if presetn assume there's an existing install.
* Otherise, check if `/opt/netdata/bin/netdata` exists, if so assume
  there's an existing install, otherwise continue under the assumption
  there isn't.
* If there is an existing install, check if it's one done through our
  install scripts, if not stop the installation process.
* If it is one of our installs, look for and attempt to use the updater
  script to update the existing install (exiting with an error message
  if that fails).

This also adds a switch called '--allow-duplicate-install' which will
cause the install to continue if it would normally refuse to continue
due to an existing install.