Trying to clean up the tester
Because the test results do not make any sense at all. Added a baseline to compare against that does not do any
replication at all. The data on the two sides does not match anywhere. There seem to be data-points on the receiver
side that do not advance the last stored time. The gaps in the collection times are different. In some places the
collected values differ for no obvious reason. There are more differences than similarities between the two sides.

Next step is to check out an old version of master, insert the debug and the tester and rerun the baseline on the
old code to verify that my last changes did not break anything. If it behaves the same way then...