Fixed interpolator sync problems
Finally found the underlying problem that was destablizing the interpolator - it was clock slew. Specifically it was
because we use the interval between a collection point and the last stored point to perform interpolation. When we
do this on the receiver side the idea is to communicate the deltas between collection points so that the sequence is
the same. BUT the interpolator is a real-time loop and it uses the current clock to drive the local stored times and
then the deltas between the stores and the collection points to do the interpolation. Obviously this cannot work
reliably because the stored times are lagged on the receiver and so even if the collection times exhibit a constant
lag the phase between the stores and the collections on the reciver is constantly changing. Somehow this was not
visible in the old streaming code - seems to be highly tuned to avoid it. I am not that smart so I have just cheated
and made sure that the interpolators calculate the same sequence on numbers on both sides of the link. To do this we use
different semantics for the deltas, we now transmit the delta between the collection point and stored time on the
sender so that we can project this into the reviever's time-domain. i.e. we just apply the delta to the local stored
time and construct a synthetic collection point on the receiver so that the phase remains constant.

Cannot test this properly yet as my docker installation has become weird and I need to clean out containers and logs
to get it back to normal. The charts look correct though regardless of if the data is streamed or replicated.