Remove runtime support for Polymorphic Linux from our Docker containers.
This removes our runtime support for Polyverse's Polymorphic Linux from
our Docker containers. The actual support is currently broken due to
upstream changes, but we have decided to just remove it instead of
fixing it due to it providing limited practical benefit to our users and
not being supported across all of our supported platforms.

We had previously removed build-time usage of this functionality for the
same reasons to reduce the final image size for our 64-bit x86 Docker
images.