Add systemd timer unit to handle auto updates.
This adds a systemd timer unit and associated service for running the
updaterscript. This allows better support for auto-updates on systems
that use systemd, removing the need for cron or special shims to handle
periodic tasks.

This will be used in preference to our existing auto-update support on
systems running systemd unless the user overrides this behavior.