Add proper handling for updater script in installers.
Our updater script has a built-in random delay of up to 1 hour when run
non-interactively to avoid a stampede effect for reconnections to
Netdata Cloud due to auto-updates. Due to an oversight in handling of
updates to existing installs in the installer scripts, this delay was
also occuring when the installer scripts were run non-interactively.
This commit fixes that so that non-interactive runs complete without a
delay.