All the two node tests are now working.
The fail on MiddleFastRestart is not a real error - the connection comes back up quickly enough that the network dumps
the buffer on the socket without a gap.

The real fail on MiddleSlowRestart was the way that the agent was killed inside docker, we have a real problem that 17
minutes of data is lost when the agent is killed so the metrics were empty on restart. The fix is to shut down the
agent grcefully via a SIGINT instead of a SIGKILL when docker kills the container.

Summary to stdout, details to logs. Cleaned up all the test vectors.

TODO:
* 3-node tests, need to exapnd the intervals but there should be about 40-50 cases
* Combinatorics for test conditions (db mode, new/old agent etc)