Cleanup and config generation.
Restructured the code separte the test driver from the specific scenario.

In order to support different configurations for each test scenario the manaually written config files are now
generated by the test driver.

Found and fixed a bug in the tester - it was comparing the wrong datasets at the end of the test.

The bug fix in the tester has unmasked a bug in the agent - need to go back into the interpolator to see why
the streaming is not replicating any more.