Remove premature optimization from the code
The sender thread avoided locking the circular buffer to check if there was outstanding data on the connection. The
condition it needs (unsent data) grows monotonically w.r.t. other threads as the collectors can add data but only
this thread can remove it. However, it cached the pointer into the buffer as a side-effect and then reused it later
during the transmission. This fails if the buffer is resized by a collector thread. Peeking at the buffer sizes without
locking could fail in the same situation.

The optimization is removed and the sender thread now locks the mutex before checking the buffer, throws away the
data buffer and releases the mutex over the poll() operation. It then reacquires the mutex and checks the buffer size
and data pointer again when it performs the send.