Unified access to node containers
The hard-coded shell strings in the test procedures did not match the recent change of project name in
docker, and some of the kill/restart steps were not using the correct signal. Pushed all of this logic
into the test state to simplify the test bodies and ensure it uses consistent naming / access.