Fix new initial conditions for replication
Replicating the initial data on a metric (instead of a gap in the middle of a series) has proven to be complex.
We need to account for metrics where the initial data-values go through the interpolator but are not stored in
the database. This creates situations where the replication window can be empty but the receiving Netdata does
not know if it will remain empty or not: the last_collected and last_stored times on the metric are both zero
so the defined time-range is indeterminate.

This has also highlighted cases that were not properly accounted for - the dimensions within a chart may start
and stop at different times so the replication window is not dense. Each row can be a different length.

The replication code now handles these different cases correctly and distributed the data-points that it receives
into the target dimensions before handling the advance of the metric window by the entire range.

Known issue: initial values are still zero in the replicated data when they are non-zero in the sending size.
This can happen both on the initial request (where there could be a one-point difference between the stored and
streamed values) and in the intermediate-gap case where the stored and streamed points must match.