Fix a race hazard between streaming and replication
The replication of a value can occur just before the value is streamed, this would cause the value to be received
again while it already in the database and run through the interpolator in the wrong state. The streaming parser
now skips such values.

The state of the interpolators do not match up after the replication. Tried sending the collection usec offset but
it is not matching on both sides.

Current issues:
   * The last value in the window becomes very small.
   * The sender database thinks there is one more value stored than has been interpolated??