Change replicated state of the chart to be midway through the done_push operation.
The fill_gaps replication happens on the collector thread at the point point that would be the done_push for normal
streaming. This avoids race hazards in the chart state between the thread doing the update and doing the replication
in another thread. However this means that the replication happens when the chart is in a slightly inconsistent state
as the last_collected_time has already been wound forwards for the update and the length of the window is stored in
the difference between the local variables now_collected_ut and last_collected_ut inside the interpolator.

Account for this difference by recalculating the old collection point and replicating that so that the interpolation
on the receiver machine will see the same sequence of timestamps.

This completes the alignment of all values shown in the exhaustive debug dump from the earlier commit but the first
value is still interpolated incorrectly. More investigation ...