Update debug: drill deeper into dbengine and remove D_RRD_CALLS to clean up log
The state dump now drills into the dbengine page cache to find the values that have been stored. The output from the
ParentShortRestart test-case shows that the query cannot retrieve data from the system.load chart, e.g.

{ "data": [ [ 1597769395, null, null, null ],
            [ 1597769390, null, null, null ],
            [ 1597769385, 2.029, 1.31, 1.12 ]
    ],

The log sequence shows replication of a single point at 1597769385 (correctly) followed by interpolation that gets
as far as "interpolate-scale" but not "interpolate-store" for the next two points. This suggests that the first loop
is scaling points but the second loop is not emitting any time points?