Alignment problem is gone
The alignment problem on system.load is gone. It was caused by the metadata reload reseting the chart times which
causes the store_next flag to get wiped and null values inserted into the database. There is a call in place that
prevents this by comparing the dimension stored times range against the chart during the reload and updating it if
the dimension is newer. Changing the debug code in that procedure (and the test length) caused the problem to
disappear. Not sure why.

There seems to be some instability in the test sequences. Currently investigating what is causing it...