Debian packaging (#8982)
* Add protobuf-compiler

This adds the package protobuf-compiler to the dependencies. Without it,
prometheus support will not be included.

* Remove libraries from dependencies

There is no need to include the libraries since they will automaticaly 
be added by ${shlibs:Depends}

* Remove various control files

- control.buster
- control.eoan
- control.xenial
Default control file can be used

- control.jessie
About to be unsupported

- control.wheezy
unsupported

- control.trusty
Since Netdata need libuv1-dev to build, it cannot be build on Trusty
anymore because it doesn't have that library.

* Fix sysv init scripts

Set the location of the pidfile using $PIDFILE in the init script and in
the defaults file. This way no changes are needed when compiling for
systemd or sysv.

* Update documentation to reflect the changes

* Add note about updating the changelog

* Update install instructions

* Remove travis control file adjustments

Since these control files have been removed, this code is now obsolete.

* Remove unused control.focal file