Update temporary directory checking in installer.
This updates the checks that our installer and updater make to choose
what to use for a temporary directory to be both more robust and more
portable.

The new checks use the following logic:

* For each directory to be checked, verify that it is both writable by
  the current user, and that the current user can execute files they
  write there.
* If `$TMPDIR` is set, preferentially use that.
* If that fials, try `/tmp`.
* If that also fails, fall back to `$PWD`.
* If all checks fail, bail early with an explanation instead of failing
  when we first tryto do things with the directory.

It also adds the same checks to the `netdata-installer.sh` script, which
was previously completely missing them.