Fix garbage values in replication
Tracked down the garbage values to a replication problem where the state of the two ends goes out of sync. This is
part of the problem, the initial setup of the query handle seems to fail and we hit "Sample mismatch" warning while
the replication request is serviced. Trying this hard abort of the replication request so that it will try again
after more data has flowed through the system.