rpm: Fix rpm build script version issues
When creating a package using this script, it's not straighforward that
the version is fetched at the "configure" stage, by "git describe" which
generates a 'lasttag-ncommits-lastcommit' format description. RPM
doesn't allow - in version, so the rpmbuild stage fails.

Add hints of actions to perform to get this script working
Use package standard functions for the message to be seen easily