Remove dots in front of "oid" field in examples
According to the docs: `Keep in mind that `snmpwalk` outputs the OIDs with a dot in front them. You should remove this dot when adding OIDs to the configuration file of this collector.` 
The examples that we provide after this message have a dot in front of the OID. This commit removes that dot.