Split window definition into a leading gap and a region with data
To support the legacy memory modes (and where their API differs from dbengine) we can't write leading gaps into the
data. The window_first parameter still controls the beginning of the window to allow the sender/receiver to sync up
on which window is being filled, but the window_first is the first sample in the chart within the window (on any
dimension). This drives how far we advance the RRD pointer in the legacy receiver.