Fix bug in the legacy metric storage.
The time2slot mapping adjusts for errors instead of reporting them. So after the handle is initialised there can be
a difference between the slot start/end and the time-stamp start/end.

The reverse mapping slot2time was hideously broken in every way. After checking that the code was dead I've rewritten
it to be correct.