Add a way to get build configuration info from the agent.
This adds a new option to the `-W` switch called 'buildinfo'. When
invoked with this argument, Netdata will print it's version, the
configure options, and a list of optional features and whether they are
enabled or not.

This is intended to serve three purposes:

* It allows developers to more quickly get an idea of how Netdata was
  built when triaging bug reports.
* It provides an easier way to validate changes to the build system that
  affect optional features during the development cycle.
* It provides an easier way to build CI workflows that validate that
  building under a given set of constraints results in a feature being
  enabled or not.

The actual implementation is a bit large but overall exceedingly simple,
consisting of a set of preprocessor directives to extract optional
feature state information from config.h and then a series of printf()
calls to actually report this info (which should end up optimized by
smart compilers due to all the arguments being compile-time constants).