Bug fixes: dbengine crash and broken socket.
Found another case where the dbengine ctx is uninitialized when the dimension is still in archive mode. This should
not happen (as the chart metadata has been received at this point) and probably indicates an underlying race hazard
in the metadata log.

The broken socket is caused by a large volume of data on reconnect (e.g. 1000 dimensions replicating the last 60s of
data). This fills the outgoing buffer. There is existing code in rrdpush that closes sockets with a full buffer, and
this was tearing down the connection.

The panic-close is disabled on connections that are replicating. I will rework this a little bit more as it could
lead to a problem where part of a REPBEGIN/REPDIM/REPEND sequence lands in the buffer before it fills, the rest is
dropped and the buffer continues to fill after the window opens in the network.