Try to update local tags during build to ensure consistent version.
By default, `git` does not update the list of tags for a repository
after it is first cloned unless you explicitly tell it to do so.

Because we use the most recent tag as the first part of our version
number, this can lead to strange ancient-looking version numbers that
are actually far more recent (for example, `v1.11.1-2915-g6106dd7`,
which is actually from v1.21.1 sources), potentially causing confusion
with respect to support.

This adds code to `netdata-installer.sh` to try to update the local repo
with the remote tags to ensure that we actually have sane version
numbers.