Use 5 digits for commit count in version number comparison.
The original code in `netdata-updater.sh` includes a hard-coded
assumption that a version number will never have more than 3 digits for
the commit count. This is of course wrong in a handful of cases, which
will then get stuck on the older version because of the effects of digit
counts on comparisons.

This updates from our current 999 commit limit to a limit of 99999
commits, which is a reasonable expectation that we should never get
_that_ far.