Removed lifecycle checks from Travis CI. (#10305)
They consistently do not catch any issues in our own code except for
syntax errors, which are already caught by shellcheck. However, they
account for a vast majority of the runtime of our Travis CI jobs for the
agent and thus are a major contributing factor to how long it takes for
checks to finish in Travis. Remove them so they stop causing us
problems.