Fixed option parsing in kickstart.sh.
The existing code is dependent on options specifically for the kickstart
script coming before any options to be passed to the installer,
resulting in some options being ignored if the wyere ordered in certain
ways.

This fixes the parsing to explicitly parse _all_ options so that we are
not dependent on option order.