Fixed handling of environment file in updater script.
Currently, when the script auto-updates, it is unable to find the
environment file in many cases due to a coding oversight.

This updates the script to pass the environment file path to the updated
copy, as well as more robustly looking for the file itself instead of
just failing if it was not substituted in during install.