Changed Docker image tagging to use semver tags for releases.
This allows users to use image names like `netdata/netdata:1` or
`netdata/netdata@1.29` and track the most up-to-date release that
matches that version prefix.

Such usage is a common practice for projects using semantic versioning
like we are.

This has a side effect, however, of remivng the `v` from the start of
our version tags. Not having it is also more consistent with how a vast
majority of other projects handle version tags, but users will need to
be notified about the change.