Fixed handling of perf.plugin capabilities.
CAP_PERFMON is only available with Linux kernel 5.8 or newer, and needs
a sufficiently up-to-date userspace to be able to set it.

With this change, we fall back to CAP_SYS_ADMIN (which CAP_PERFMON is a
proper subset of) if we cannot set CAP_PERFMON for the perf plugin.