Allow arbitrary options to be passed to make from netdata-installer.sh.
THis is mostly intended to allow users to override the number of
parallel jobs used during the build process. It’s been made generic as
a handful of other options (most notably `--load-average`) are also
potentially useful and because the generic behavior is consistent with
how most source-based distros handle this type of thing.